separate colored_logger
